Updated dotfiles
- making find posix compatible
- using os for lf
- do not install nvim.appimage on other systems then linux

+++ shell/common/nredf_functions.sh
@@ -68,6 +68,9 @@ function _nredf_set_defaults() {
 }
 
 function _nredf_install_nvim() {
+  _nredf_get_sys_info
+
+  [[ "${OS}" != "linux" ]] && return 1
   if [[ ! -f "${HOME}/.local/bin/nvim" ]] || [[ "$(${HOME}/.local/bin/nvim --version | head -1)" != "$(curl -sH 'Accept: application/vnd.github.v3+json' https://api.github.com/repos/neovim/neovim/releases/tags/nightly | grep -Po '"name":"\K.*?(?=")' | head -1)" ]]; then
     echo -e '\033[1mDownloading neovim\033[0m'
     [[ -d "${HOME}/.cache/vim/squashfs-root" ]] && rm -rf "${HOME}/.cache/vim/squashfs-root"
@@ -88,7 +91,7 @@ function _nredf_install_lf() {
 
   if [[ ! -f "${HOME}/.local/bin/lf" ]] || [[ "$(_nredf_github_latest_release gokcehan lf)" != "$(${HOME}/.local/bin/lf -version)" ]]; then
     echo -e '\033[1mInstalling lf\033[0m'
-    curl -Ls "https://github.com/gokcehan/lf/releases/latest/download/lf-linux-${ARCH}.tar.gz" | tar xzf - -C ${HOME}/.local/bin/
+    curl -Ls "https://github.com/gokcehan/lf/releases/latest/download/lf-${OS}-${ARCH}.tar.gz" | tar xzf - -C ${HOME}/.local/bin/
     chmod +x ${HOME}/.local/bin/lf
   fi
 }
